// ui/src/lib/api/model/internalProfilesProfile.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { Browser } from './browser';
import type { InternalProfilesStorageBackend } from './internalProfilesStorageBackend';

export interface InternalProfilesProfile {
  /** Computed fields */
  active_sessions?: number;
  browser?: Browser;
  created_at?: string;
  description?: string;
  id?: string;
  last_used_at?: string;
  metadata?: number[];
  name?: string;
  size_bytes?: number;
  storage_backend?: InternalProfilesStorageBackend;
  total_sessions?: number;
  updated_at?: string;
}