// ui/src/lib/api/model/internalProfilesCreateProfileRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { InternalProfilesCreateProfileRequestBrowser } from './internalProfilesCreateProfileRequestBrowser';

export interface InternalProfilesCreateProfileRequest {
  browser: InternalProfilesCreateProfileRequestBrowser;
  /** @maxLength 1000 */
  description?: string;
  /**
   * @minLength 1
   * @maxLength 255
   */
  name: string;
}