// ui/src/lib/api/model/internalProfilesCreateProfileRequestBrowser.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

export type InternalProfilesCreateProfileRequestBrowser = typeof InternalProfilesCreateProfileRequestBrowser[keyof typeof InternalProfilesCreateProfileRequestBrowser];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InternalProfilesCreateProfileRequestBrowser = {
  chrome: 'chrome',
  chromium: 'chromium',
  firefox: 'firefox',
  edge: 'edge',
  webkit: 'webkit',
  safari: 'safari',
} as const;